backend/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.13
decorator==5.2.1
distlib==0.3.9
dj-database-url==3.0.1
Django==4.2.16
django-admin-rangefilter==0.13.2
django-cors-headers==4.4.0
django-ratelimit==4.1.0
django-redis==6.0.0
djangorestframework==3.15.2
djangorestframework-csv==3.0.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
executing==2.2.0
filelock==3.18.0
gunicorn==21.2.0
inflection==0.5.1
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
pandas==2.3.1
parso==0.8.4
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
uritemplate==4.2.0
virtualenv==20.30.0
wcwidth==0.2.13
whitenoise==6.7.0
